# Remove JsonReaderOption to track position by UTF-8 code points