Essentially reverts https://github.com/dotnet/corefxlab/pull/2540.

Gain back ~20%.

``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-009701
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.3 (CoreCLR 4.6.26725.06, CoreFX 4.6.26725.05), 64bit RyuJIT

```

**Before:**

|                               Method | IsDataCompact |   TestCase |           Mean |         Error |        StdDev |         Median | Scaled | ScaledSD | Allocated |
|------------------------------------- |-------------- |----------- |---------------:|--------------:|--------------:|---------------:|-------:|---------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |       **606.9 ns** |     **12.031 ns** |     **15.215 ns** |       **614.6 ns** |   **0.82** |     **0.03** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True |  BasicJson |       742.5 ns |     14.832 ns |     23.525 ns |       743.6 ns |   1.00 |     0.00 |       0 B |
|                                      |               |            |                |               |               |                |        |          |           |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |       **112.0 ns** |      **1.779 ns** |      **1.664 ns** |       **111.9 ns** |   **1.06** |     **0.04** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True | HelloWorld |       106.0 ns |      2.132 ns |      4.057 ns |       104.8 ns |   1.00 |     0.00 |       0 B |
|                                      |               |            |                |               |               |                |        |          |           |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |       **979.2 ns** |     **19.473 ns** |     **53.307 ns** |       **961.1 ns** |   **0.75** |     **0.04** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True |   Json400B |     1,313.1 ns |     25.600 ns |     25.142 ns |     1,311.0 ns |   1.00 |     0.00 |       0 B |
|                                      |               |            |                |               |               |                |        |          |           |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** |   **994,164.2 ns** | **20,174.711 ns** | **28,282.127 ns** |   **982,818.3 ns** |   **0.91** |     **0.03** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True |  Json400KB | 1,095,538.9 ns |  8,554.799 ns |  7,583.606 ns | 1,095,355.6 ns |   1.00 |     0.00 |       0 B |
|                                      |               |            |                |               |               |                |        |          |           |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |    **92,671.9 ns** |  **1,332.915 ns** |  **1,113.044 ns** |    **92,388.6 ns** |   **0.80** |     **0.07** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True |   Json40KB |   116,364.3 ns |  3,623.939 ns | 10,685.267 ns |   113,176.6 ns |   1.00 |     0.00 |       0 B |
|                                      |               |            |                |               |               |                |        |          |           |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |     **7,224.4 ns** |    **113.743 ns** |     **94.981 ns** |     **7,256.9 ns** |   **0.68** |     **0.01** |       **0 B** |
|              ReaderUtf8JsonEmptyLoop |          True |    Json4KB |    10,628.9 ns |    176.690 ns |    156.631 ns |    10,570.9 ns |   1.00 |     0.00 |       0 B |

**Now:**

|                               Method | IsDataCompact |   TestCase |          Mean |        Error |       StdDev | Allocated |
|------------------------------------- |-------------- |----------- |--------------:|-------------:|-------------:|----------:|
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **BasicJson** |     **505.52 ns** |    **11.100 ns** |    **13.214 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** | **HelloWorld** |      **94.34 ns** |     **1.657 ns** |     **1.469 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json400B** |     **762.27 ns** |     **5.404 ns** |     **4.791 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |  **Json400KB** | **827,304.21 ns** | **3,062.236 ns** | **2,390.793 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |   **Json40KB** |  **77,472.61 ns** |   **390.856 ns** |   **365.607 ns** |       **0 B** |
| **ReaderSystemTextJsonLabSpanEmptyLoop** |          **True** |    **Json4KB** |   **5,831.29 ns** |   **115.613 ns** |   **133.140 ns** |       **0 B** |